optimize auth action&reducer code, no new function/feature
