+++ src/EventHive/Auth/LogIn/index.js
@@ -2,7 +2,7 @@ import React, { Fragment, useState } from 'react';
 import { Link, Navigate } from 'react-router-dom';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import { login } from '../../../actions/auth';
+import { login } from '../client';
 
 const Login = ({ login, isAuthenticated }) => {
 	const [formData, setFormData] = useState({
