+++ src/EventHive/Auth/client.js
@@ -1,21 +1,26 @@
-import axios from '../config/axiosConfig';
+import axios from '../../config/axiosConfig';
 
-import { showAlert } from '../utils/alertHelper';
+import { showAlert } from '../../utils/alertHelper';
+import { userLoaded, userAuthenticated, userNotAuthenticated } from './AuthReducer';
 
 export const loadUser = () => async (dispatch) => {
+	console.log("begin loadUser()============");
 
 	if (localStorage.apiKey) {
 		axios.defaults.headers.common['x-api-key'] = localStorage.apiKey;
 	} else {
 		delete axios.defaults.headers.common['x-api-key'];
 	}
-
+	
 	try {
 		const res = await axios.get('/users/load');
 
-		dispatch({ type: 'USER_LOADED', payload: res.data });
+		// dispatch({ type: 'USER_LOADED', payload: res.data });
+		dispatch(userLoaded(res.data));
+
 	} catch (err) {
-		dispatch({ type: 'NotAuthenticated' });
+		// dispatch({ type: 'NotAuthenticated' });
+		dispatch(userNotAuthenticated());
 	}
 };
 
@@ -32,11 +37,17 @@ export const register =
 		try {
 			const res = await axios.post('/users/register', body, config);
 
-			dispatch({
-				type: 'Authenticated',
-				payload: res.data,
-			});
+			// dispatch({
+			// 	type: 'Authenticated',
+			// 	payload: res.data,
+			// });
+			// dispatch(loadUser());
+
+
+			dispatch(userAuthenticated(res.data));
 			dispatch(loadUser());
+
+
 			showAlert('success', 'Success!', 'SignUp Successfully');
 		} catch (err) {
 			const errors = err.response.data.errors;
@@ -49,9 +60,11 @@ export const register =
 				);
 			}
 
-			dispatch({
-				type: 'NotAuthenticated',
-			});
+			// dispatch({
+			// 	type: 'NotAuthenticated',
+			// });
+
+			dispatch(userNotAuthenticated());
 		}
 	};
 
@@ -67,11 +80,14 @@ export const login = (email, password, isAdministrator) => async (dispatch) => {
 
 	try {
 		const res = await axios.post('/users/login', body, config);
-		dispatch({
-			type: 'Authenticated',
-			payload: res.data,
-		});
+		// dispatch({
+		// 	type: 'Authenticated',
+		// 	payload: res.data,
+		// });
 
+		// dispatch(loadUser());
+
+		dispatch(userAuthenticated(res.data));
 		dispatch(loadUser());
 
 		showAlert('success', 'Success!', 'Login Successfully');
@@ -85,13 +101,15 @@ export const login = (email, password, isAdministrator) => async (dispatch) => {
 				errors.map((error) => error.msg).join('<br/>')
 			);
 		}
-		dispatch({
-			type: 'NotAuthenticated',
-		});
+		// dispatch({
+		// 	type: 'NotAuthenticated',
+		// });
+		dispatch(userNotAuthenticated());
 	}
 };
 
 export const logout = () => (dispatch) => {
-	dispatch({ type: 'NotAuthenticated' });
+	// dispatch({ type: 'NotAuthenticated' });
+	dispatch(userNotAuthenticated());
 	showAlert('success', 'Success!', 'Logout Successfully');
 };
