+++ src/reducers/index.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import auth from './auth';
+import auth from '../EventHive/Auth/AuthReducer';
 import EventsReducer from '../components/Events/EventsReducer';
 import ProfileReducer from '../EventHive/Profile/ProfileReducer';
